requirements.txt
# Core framework
Django==5.1.4

# REST API
djangorestframework==3.15.2

# HTTP requests (used for X/Twitter API notifications)
requests==2.32.3

# MySQL database driver
mysqlclient==2.2.4

# Documentation
Sphinx==8.1.3
sphinx-rtd-theme==3.0.2
